使用map优化了性能
使用map优化了性能，通过了大数据集的测试

+++ Query.cpp
@@ -78,7 +78,7 @@ void Query::execute()
 	{
 		ifstream posReader((inputDir.getPathName() + "/pos.dict").c_str());
 		int termid;
-		long long pos;
+		long pos;
 		
 		while (posReader >> termid >> pos) {
 			posDict[termid] = pos;
